feat(returns): supply non-returnable model change as a trait
Ship NonReturnableProductTrait (non_returnable column + accessors) so an
application enables the flag by using the trait and implementing
NonReturnableProductInterface. Update the install instructions and the test-app
Product accordingly. Exclude the downstream-consumed trait from PHPStan's src
scope to avoid a false-positive trait.unused.

Refs #18
